android-app-testing/requirements.txt
Appium-Python-Client==5.3.0
selenium==4.42.0
requests==2.32.4
typing_extensions==4.15.0
pydantic==2.11.7
fastapi==0.116.1
uvicorn==0.35.0
python-dotenv==1.1.1
PyYAML==6.0.2
colorama==0.4.6
rich==14.3.3
psutil==6.1.1
